requirementsOUTDATED.txt
discord.py
openai
ffmpeg-python
nltk
youtube_dl
pinecone-sdk
azure-storage-blob